Add error handling for binding the UDP socket
Instead of just leaving a comment.

I also changed the last argument of ConnectToAddress
to be 1 so that the call is synchronous. Given that
we are working with UDP sockets, there is no actual
connection, we are just setting some stuff in the
kernel to simplify sending packets in the future,
so the call should only fail if we configure something
incorrectly and take much less than 1 second.