Create and expose our own concurrent queue
Otherwise dispatch_barrier_sync does not work. This
is still something that I need to look into once
the dev forums are up again. I found nothing in the
documentation that says that I cannot barrier_sync
on global queues, but I could see reasons why it
would not be allowed.