Update tests for graylog_queue() changes
And also implement more of the unimplemented tests
so that I can at least toy with the super
concurrency of the logger.

The tests are still disabled at the moment as they
cause some serious timing issues because they
finish before graylog can finish parsing all of
the messages that were sent to it, which causes
messages to spill over into other tests.